+++ tobis-space/src/App.tsx
@@ -6,6 +6,7 @@ import BoardGameCommunity from './pages/BoardGameCommunity'
 import BoardGameRules from './pages/BoardGameRules'
 import BoardGameUpdates from './pages/BoardGameUpdates'
 import Stories from './pages/Stories'
+import StoryLanding from './pages/StoryLanding'
 import Chapter from './pages/Chapter'
 import Drawings from './pages/Drawings'
 import CheckoutSuccess from './pages/CheckoutSuccess'
@@ -22,6 +23,7 @@ export default function App() {
           <Route path="updates" element={<BoardGameUpdates />} />
         </Route>
         <Route path="stories" element={<Stories />}>
+          <Route index element={<StoryLanding />} />
           <Route path=":chapterSlug" element={<Chapter />} />
         </Route>
         <Route path="drawings" element={<Drawings />} />
